+++ LedStripBase.cpp
@@ -22,11 +22,11 @@
 
 #include "LedStripBase.h"
 
-BaseLedStrip::BaseLedStrip(uint8_t ledCount) {
+BaseLedStrip::BaseLedStrip(uint16_t ledCount) {
     this->ledCount = ledCount;
 }
 
-BaseLedStrip::BaseLedStrip(uint8_t ledCount, BaseLedStripAdapter* adapter) {
+BaseLedStrip::BaseLedStrip(uint16_t ledCount, BaseLedStripAdapter* adapter) {
     this->ledCount = ledCount;
     this->adapter = adapter;
 }
@@ -36,11 +36,11 @@ BaseLedStrip::~BaseLedStrip() {
     delete &this->adapter;
 }
 
-uint8_t BaseLedStrip::getLedCount() {
+uint16_t BaseLedStrip::getLedCount() {
     return this->ledCount;
 }
 
-void BaseLedStrip::setLedCount(uint8_t ledCount) {
+void BaseLedStrip::setLedCount(uint16_t ledCount) {
     // Set the LED count field
     this->ledCount = ledCount;
 
@@ -56,50 +56,50 @@ void BaseLedStrip::setAdapter(BaseLedStripAdapter* adapter) {
     this->adapter = adapter;
 }
 
-LedStripColor BaseLedStrip::getLedColor(uint8_t ledIndex) {
+LedStripColor BaseLedStrip::getLedColor(uint16_t ledIndex) {
     return (*this->adapter).getLedColor(ledIndex);
 }
 
-void BaseLedStrip::setLedColor(uint8_t ledIndex, LedStripColor color) {
+void BaseLedStrip::setLedColor(uint16_t ledIndex, LedStripColor color) {
     (*this->adapter).setLedColor(ledIndex, color);
 }
 
-void BaseLedStrip::setLedColor(uint8_t ledIndex, uint8_t redChannel) {
+void BaseLedStrip::setLedColor(uint16_t ledIndex, uint8_t redChannel) {
     (*this->adapter).setLedColor(ledIndex, redChannel);
 }
 
-void BaseLedStrip::setLedColor(uint8_t ledIndex, uint8_t redChannel, uint8_t greenChannel) {
+void BaseLedStrip::setLedColor(uint16_t ledIndex, uint8_t redChannel, uint8_t greenChannel) {
     (*this->adapter).setLedColor(ledIndex, redChannel, greenChannel);
 }
 
-void BaseLedStrip::setLedColor(uint8_t ledIndex, uint8_t redChannel, uint8_t greenChannel, uint8_t blueChannel) {
+void BaseLedStrip::setLedColor(uint16_t ledIndex, uint8_t redChannel, uint8_t greenChannel, uint8_t blueChannel) {
     (*this->adapter).setLedColor(ledIndex, redChannel, greenChannel, blueChannel);
 }
 
-void BaseLedStrip::setLedColor(uint8_t ledIndex, uint8_t redChannel, uint8_t greenChannel, uint8_t blueChannel,
+void BaseLedStrip::setLedColor(uint16_t ledIndex, uint8_t redChannel, uint8_t greenChannel, uint8_t blueChannel,
                                uint8_t alphaChannel) {
     (*this->adapter).setLedColor(ledIndex, redChannel, greenChannel, blueChannel, alphaChannel);
 }
 
-void BaseLedStrip::setRangeLedColors(uint8_t fromLedIndex, uint8_t toLedIndex, LedStripColor color) {
+void BaseLedStrip::setRangeLedColors(uint16_t fromLedIndex, uint16_t toLedIndex, LedStripColor color) {
     (*this->adapter).setRangeLedColors(fromLedIndex, toLedIndex, color);
 }
 
-void BaseLedStrip::setRangeLedColors(uint8_t fromLedIndex, uint8_t toLedIndex, uint8_t redChannel) {
+void BaseLedStrip::setRangeLedColors(uint16_t fromLedIndex, uint16_t toLedIndex, uint8_t redChannel) {
     (*this->adapter).setRangeLedColors(fromLedIndex, toLedIndex, redChannel);
 }
 
-void BaseLedStrip::setRangeLedColors(uint8_t fromLedIndex, uint8_t toLedIndex, uint8_t redChannel,
+void BaseLedStrip::setRangeLedColors(uint16_t fromLedIndex, uint16_t toLedIndex, uint8_t redChannel,
                                      uint8_t greenChannel) {
     (*this->adapter).setRangeLedColors(fromLedIndex, toLedIndex, redChannel, greenChannel);
 }
 
-void BaseLedStrip::setRangeLedColors(uint8_t fromLedIndex, uint8_t toLedIndex, uint8_t redChannel, uint8_t greenChannel,
+void BaseLedStrip::setRangeLedColors(uint16_t fromLedIndex, uint16_t toLedIndex, uint8_t redChannel, uint8_t greenChannel,
                                      uint8_t blueChannel) {
     (*this->adapter).setRangeLedColors(fromLedIndex, toLedIndex, redChannel, greenChannel, blueChannel);
 }
 
-void BaseLedStrip::setRangeLedColors(uint8_t fromLedIndex, uint8_t toLedIndex, uint8_t redChannel, uint8_t greenChannel,
+void BaseLedStrip::setRangeLedColors(uint16_t fromLedIndex, uint16_t toLedIndex, uint8_t redChannel, uint8_t greenChannel,
                                      uint8_t blueChannel, uint8_t alphaChannel) {
     (*this->adapter).setRangeLedColors(fromLedIndex, toLedIndex, redChannel, greenChannel, blueChannel, alphaChannel);
 }
